Refactor: Implement RFID Repository and Update Handler
This commit introduces a dedicated RFIDRepository to handle student data retrieval by RFID, improving the application's architecture and separation of concerns.
